Cache which adapter class handles each item class
